// src/lib/api/generated/models/v1OTCStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1OTCStatus {
    OTC_STATUS_UNSPECIFIED = 'OTC_STATUS_UNSPECIFIED',
    OTC_STATUS_OPEN = 'OTC_STATUS_OPEN',
    OTC_STATUS_SUPERSEDED = 'OTC_STATUS_SUPERSEDED',
    OTC_STATUS_ACCEPTED = 'OTC_STATUS_ACCEPTED',
    OTC_STATUS_WITHDRAWN = 'OTC_STATUS_WITHDRAWN',
    OTC_STATUS_EXPIRED = 'OTC_STATUS_EXPIRED',
    OTC_STATUS_CANCELLED = 'OTC_STATUS_CANCELLED',
    OTC_STATUS_REJECTED = 'OTC_STATUS_REJECTED',
}